+++ test/simple_oauth/header_body_hash_digest_test.rb
@@ -10,6 +10,10 @@ module SimpleOAuth
 
     BODY = '{"status":"testing"}'.freeze
 
+    def setup
+      Signature.reset!
+    end
+
     def teardown
       Signature.reset!
     end